streaming/requirements_streaming.txt
pyspark==3.5.0
kafka-python==2.0.2
minio==7.2.0
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2